add check command mv

+++ app/main.py
@@ -7,12 +7,14 @@ def move_file(command_line: str) -> None:
     second_file_path = second_file[:last_slash_index]
     content = ""
 
-    if not os.path.exists(second_file_path):
-        os.makedirs(second_file_path)
+    if command == "mv" and last_slash_index != "-1":
+        if not os.path.exists(second_file_path):
+            os.makedirs(second_file_path)
 
-    with open(first_file, "r") as file_content:
-        content += file_content.read()
+        with open(first_file, "r") as file_content:
+            content += file_content.read()
 
-    with open(second_file, "w") as file_out:
-        file_out.write(content)
         os.remove(first_file)
+
+        with open(second_file, "w") as file_out:
+            file_out.write(content)
